IOS-3056 Add EthereumNetworkProvider
